Refactor panda/board/jungle/stm32*/*.h
