+++ board/jungle/stm32f4/board.h
@@ -1,7 +1,5 @@
-#include "boards/board_declarations.h"
+#pragma once
+#include "boards/board.h"
 #include "boards/board_v1.h"
 
-void detect_board_type(void) {
-  hw_type = HW_TYPE_V1;
-  current_board = &board_v1;
-}
+void detect_board_type(void);
