+++ board/jungle/stm32f4/src/board.c
@@ -1,0 +1,6 @@
+#include "../board.h"
+
+void detect_board_type(void) {
+  hw_type = HW_TYPE_V1;
+  current_board = &board_v1;
+}
