+++ board/jungle/stm32h7/board.h
@@ -1,9 +1,7 @@
-#include "boards/board_declarations.h"
+#pragma once
+#include "boards/board.h"
 
 #include "stm32h7/lladc.h"
 #include "boards/board_v2.h"
 
-void detect_board_type(void) {
-  hw_type = HW_TYPE_V2;
-  current_board = &board_v2;
-}
+void detect_board_type(void);
